# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a insiderGo task project for managing messages.",
        "title": "insiderGo Project",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8090",
    "basePath": "/",
    "paths": {
        "/generate-token": {
            "post": {
                "description": "Generates a JWT token for testing or authentication",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Generate JWT Token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.TokenResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/instance/unsent-messages": {
            "get": {
                "description": "Retrieve a list of unsent messages from the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instance"
                ],
                "summary": "Get unsent messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "\u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.MessageDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/task/start": {
            "post": {
                "description": "Starts the periodic task service to process tasks at regular intervals",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PeriodicTask"
                ],
                "summary": "Start the periodic task service",
                "parameters": [
                    {
                        "type": "string",
                        "description": "\u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/task/stop": {
            "post": {
                "description": "Stops the periodic task service and halts task processing",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PeriodicTask"
                ],
                "summary": "Stop the periodic task service",
                "parameters": [
                    {
                        "type": "string",
                        "description": "\u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/temp/getAllList": {
            "get": {
                "description": "Retrieve data from TempMemory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TempMemory"
                ],
                "summary": "Get keys with pagination (GET)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "\u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/temp/getList": {
            "post": {
                "description": "Retrieve data from TempMemory using skip and limit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TempMemory"
                ],
                "summary": "Get keys with pagination (POST)",
                "parameters": [
                    {
                        "description": "Pagination parameters",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TempListPayload"
                        }
                    },
                    {
                        "type": "string",
                        "description": "\u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/unsent-messages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all messages with a sending status of \"not_sent\"",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DB"
                ],
                "summary": "Fetch unsent messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "\u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.MessageDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dtos.MessageDTO": {
            "type": "object",
            "properties": {
                "message_content": {
                    "type": "string"
                },
                "recipient_phone_number": {
                    "type": "string"
                },
                "sending_status": {
                    "type": "string"
                }
            }
        },
        "dtos.TokenResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "model.TempListPayload": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "skip": {
                    "type": "integer"
                }
            }
        }
    }
}